requirements.txt
requests
bs4
datetime
kaggle
logging
python-dotenv
